Move test harness to its own module and extend with timeout helpers
